asr/requirements.txt
openai-whisper
soundfile
librosa
transformers
numpy
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.3.0+cu118
torchaudio==2.3.0+cu118
jiwer
fastapi==0.115.12
uvicorn[standard]==0.34.2
 
